etc:함수 명 변경 디렉토리 명 변경 및 함수 파일 위치 변경
<description>
함수 명 변경(ft_get_command -> ft_handle_command / ft_handle_parsing ->
ft_route_command)
디렉토리 명 변경(./src/parsing->./src/commnad)
함수(ft_handle_command)위치 변경(./src/main.c ->
./src/command/ft_handle_command.c)
<issue-number>
