+++ include/minishell.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minishell.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dakim <dakim@student.42seoul.kr>           +#+  +:+       +#+        */
+/*   By: dakim <dakim@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/10 16:30:38 by seunkim           #+#    #+#             */
-/*   Updated: 2020/08/13 18:34:30 by dakim            ###   ########.fr       */
+/*   Updated: 2020/08/15 13:54:30 by dakim            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -61,7 +61,9 @@
 
 int		get_next_line(int fd, char **line);
 char	*ft_trim_str(const char *str);
-int		ft_handle_parsing(const char *str);
+
+void	ft_handle_command(void);
+int		ft_route_command(const char *str);
 
 int		ft_handle_error(const int error_num, const int exit_num,
 							void *custom_error, void *content);
