+++ src/command/ft_route_command.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_handle_parsing.c                                :+:      :+:    :+:   */
+/*   ft_route_command.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dakim <dakim@student.42seoul.kr>           +#+  +:+       +#+        */
+/*   By: dakim <dakim@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/12 15:27:38 by dakim             #+#    #+#             */
-/*   Updated: 2020/08/13 18:13:47 by dakim            ###   ########.fr       */
+/*   Updated: 2020/08/15 13:53:51 by dakim            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ char		*ft_trim_str(const char *str)
 	return (return_str);
 }
 
-int			ft_handle_parsing(const char *str)
+int			ft_route_command(const char *str)
 {
 	if (ft_strnstr(str, ECHO_STR, ft_strlen(ECHO_STR)))
 		return (ECHO_INT);
